services/auth-svc/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
redis[hiredis]==5.2.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
